This should fix the problem with splitter positions not being stored when the workspace their on isn't viewed before closing GCS, as Edouard found out.
The source of this problem is that QSplitter.sizes() only returns the sizes set by setSizes() after the QSplitter is actually shown. Even though save/restoreState manages to retrieve them in the same situation. I think it's a bug, not sure if I can be motivated enough to report it though :)


git-svn-id: svn://svn.openpilot.org/OpenPilot/trunk@1699 ebee16cc-31ac-478f-84a7-5cbb03baadba