Experimental Branch I added an experimental branch.  This is the location to put all code that is too hot for trunk.
In order to prevent the need to back- or sideport changes on trunk here, its
not a complete copy of the repository but just the experimental modules itself.

In order to build and "use" them its therefore necessary to copy them into the
respective locations of the users local trunk checkout.



git-svn-id: svn://svn.openpilot.org/OpenPilot/branches/Experimental@1736 ebee16cc-31ac-478f-84a7-5cbb03baadba